Extract src/taskui.ts — the tasks panel and the Run picker
Both task surfaces out of main.ts: the project tasks panel (pin / hide / create /
edit / delete / override) and the ▶ Run picker, with their own dialog state and
click/key handlers. 504 lines; main.ts 3,164 -> 2,721.

They ship together because they are two views of one thing — the Runnables a
project declares — and share the rule that shapes both: what can't run says so, a
blocked row greyed rather than dropped. Discovery, preferences and the dependency
chain are all in ./tasks now, so this is only the UI over them.

Eight-entry host object, same shape as settings.ts: launchTask, openInputPrompt,
handToTerminal, activeProjectCtx, activeCwd, setActive, renderAll, closePalette.
setRunner moved into ./tasks to join the map it writes.

Two boundary mistakes the guards and tsc caught, both worth knowing for the
remaining dialogs:

- The section comments lie about extent. handToTerminal, runGit and launchShell
  sit physically *between* the picker's functions and the picker's DOM handlers,
  so "from the ▶ Run picker header to the next header" swallowed three PTY
  functions. The move is two ranges, not one.
- Rewriting call sites by regex also renamed a *declaration*: handToTerminal is
  defined inside that range, and hostify turned it into `async function
  host.handToTerminal(...)`. It now skips declaration lines.

I also ran a verification diff that silently compared two empty files — the awk
range matched nothing. Redone properly: 448 vs 449 lines, differing only by a
trailing newline. Worth stating because "IDENTICAL" from a check that examined
nothing is worse than no check at all.

Smoked through the app's own controls: ⌘K still lists both task entries, and ▶
Run with nothing on stage reaches host.activeProjectCtx and toasts "No active
project". No JS errors.
